Fix F5 BIG-IP ASM plugin TypeError by correcting matchCookie usage
The F5 BIG-IP ASM plugin was incorrectly calling `matchCookie` with a tuple of
two regex patterns:

    self.matchCookie((r'^TS[a-fA-F0-9]{8}$', r'.+'))

This caused a `TypeError: first argument must be string or compiled pattern`
because `re.search()` was receiving a tuple instead of a string.

`matchCookie` is designed to take a single regex that matches the full
Set-Cookie value. Updated the plugin to use one combined regex pattern for
cookie name and value:

    self.matchCookie(r'^TS[a-fA-F0-9]{8}=.+')
    self.matchCookie(r'^TS[a-fA-F0-9]{6}=.+')

This resolves the TypeError and properly detects F5 ASM cookies.